+++ api/src/app.js
@@ -8,39 +8,9 @@ const cors = require('cors');
 const helmet = require('helmet');
 const routes = require('./routes');
 
-const http = require('http');
-const fs = require('fs');
-
-const multer = require('multer');
-const csv = require('fast-csv');
-
-const Router = express.Router;
-
-const upload = multer({ dest: 'tmp/csv/' });
-const app = express();
-const router = new Router();
-const server = http.createServer(app);
-const port = 9000;
-
-router.post('/', upload.single('file'), function (req, res) {
-  const fileRows = [];
-
-  // open uploaded file
-  csv
-    .fromPath(req.file.path)
-    .on('data', function (data) {
-      fileRows.push(data); // push each row
-    })
-    .on('end', function () {
-      console.log(fileRows);
-      fs.unlinkSync(req.file.path); // remove temp file
-      //process "fileRows" and respond
-    });
-});
-
-app.use('/upload-csv', Router);
-
 dotenv.config();
+
+const app = express();
 
 // connect to MongoDB
 mongoose
@@ -63,11 +33,3 @@ app.use(helmet());
 app.use('/', routes);
 
 module.exports = app;
-
-function startServer() {
-  server.listen(port, function () {
-    console.log('Express server listening on ', port);
-  });
-}
-
-setImmediate(startServer);
